Remove&Fix: google_calendar_controllerを削除、GoogleCalendarの認証コードをモデルに切り出す修正
