--- app/controllers/google_calendar_controller.rb
@@ -1,2 +1,0 @@
-class GoogleCalendarController < ApplicationController
-end
