--- app/helpers/google_calendar_helper.rb
@@ -1,2 +1,0 @@
-module GoogleCalendarHelper
-end
